requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.2.3
cryptography==44.0.1
python-dotenv==1.0.0
WTForms==3.0.1
email-validator==2.1.1
jira==3.8.0
urllib3==2.5.0
requests==2.32.3

# Authentication providers
ldap3==2.9.1
python-ldap==3.4.4
Authlib==1.3.0
python-keycloak==4.2.2